redbird v0.4.0 mix redbird.delete_all_sessions

Deletes all redbird sessions.

mix redbird.delete_all_sessions my_app

The first argument is the app specific key_namespace you set in your plug session config. If no argument is given, it will delete all redbird sessions.

Functions

delete_all_sessions(namespace)

run(args)

A task needs to implement run which receives a list of command line args.

Callback implementation for Mix.Task.run/1.