redbird v0.2.0 Plug.Session.REDIS

Stores the session in a redis store.

Functions

delete(conn, redis_key, kinit_options)

Removes the session associated with given session id from the store.

Callback implementation for Plug.Session.Store.delete/3.

get(conn, namespaced_key, init_options)

Parses the given cookie.

Returns a session id and the session contents. The session id is any value that can be used to identify the session by the store.

The session id may be nil in case the cookie does not identify any value in the store. The session contents must be a map.

Callback implementation for Plug.Session.Store.get/3.

init(opts)

Initializes the store.

The options returned from this function will be given to get/3, put/4 and delete/3.

Callback implementation for Plug.Session.Store.init/1.

namespace()

put(conn, redis_key, data, init_options)

Stores the session associated with given session id.

If nil is given as id, a new session id should be generated and returned.

Callback implementation for Plug.Session.Store.put/4.

redbird_namespace()